Dev viguard connection v2 oke

+++ app/Http/Livewire/Contact/Delete.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Livewire\Contact;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use App\Models\Client;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
     public $modalDeleteVisible = false;
     public $contactId;
     public $contact;
@@ -29,6 +31,7 @@ class Delete extends Component
 
     public function delete()
     {
+        $this->authorize('DELETE_RESOURCE', $this->contact->user_id);
         if ($this->contact) {
             $this->contact->delete();
         }
